TERRAFORM/modules/Dataflow/requirements.txt
apache-beam[gcp]
google-cloud-pubsub
google-cloud-bigquery
google-cloud-storage
google-api-python-client
grpcio
protobuf
argparse
logging